backend/requirements.txt
asgiref==3.5.0
backports.entry-points-selectable==1.1.1
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
cycler==0.11.0
distlib==0.3.6
filelock==3.12.2
flake8==6.0.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
fonttools==4.28.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.3.2
MarkupSafe==2.1.3
matplotlib==3.5.0
mccabe==0.7.0
numpy==1.24.3
packaging==21.3
pandas==2.0.2
pipenv==2023.6.12
platformdirs==3.5.3
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.1
setuptools-scm==6.3.2
six==1.16.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
tomli==1.2.2
typing_extensions==4.6.3
tzdata==2023.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
Werkzeug==2.3.6
